+++ sleplet/plotting/arbitrary/south_america/denoising_slepian_functions_south_america.py
@@ -17,7 +17,7 @@ NORMALISE = False
 SNR_IN = -10
 
 
-def main(snr: int, sigma: int) -> None:
+def main(snr: float, sigma: int) -> None:
     """
     denoising demo using Slepian wavelets
     """
